integrity_service/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
redis
requests
PyYAML
tenacity
